Updates the example app to use CDTReplicators properly.
The CDTTodoReplicator keeps a strong reference to the
CDTReplicator objects that it creates and sets itself as
their delegate.

The polling block has been removed.

The CDTTodoReplicator now uses the new CDTPush/PullReplication
configuration and the CDTReplicatorFactor -oneWay:error method
to build CDTReplicator objects.

Additionally, the CDTAppDelegate now keeps a persistent
CDTTodoReplicator object, which the CDTViewController uses to
fire off synchronization.